refactor: moved computed history data to pageview
