+++ web/src/pages/ProfilePage/ProfilePage.js
@@ -11,34 +11,6 @@ import ProfilePageView from "./ProfilePageView";
 import { UserContext } from "../../App";
 import { getCompetencesWithCategoryNames } from "../../utils";
 
-const HISTORY_TYPE = {
-  education: "Education",
-  work: "Work",
-};
-
-/**
- * A function that produces the props for using HistoryList component
- *
- * @param {Array<object>} historyItems - Array of history items
- * @param {*} type - type of history items
- * @returns {object}
- */
-const getHistoryProps = (historyItems = [], type) => {
-  return {
-    title: `${type} History`,
-    noItemDescription: `No ${type} History Provided`,
-    historyItems: historyItems.map((historyItem) => ({
-      id: historyItem.id,
-      organisation:
-        historyItem[type === HISTORY_TYPE.education ? "school" : "company"],
-      title:
-        historyItem[type === HISTORY_TYPE.education ? "degree" : "position"],
-      description: historyItem.description,
-      start_date: historyItem.start_date,
-      end_date: historyItem.end_date,
-    })),
-  };
-};
 
 function ProfilePage({ id, onSave }) {
   let history = useHistory();
@@ -63,24 +35,6 @@ function ProfilePage({ id, onSave }) {
     } 
   }, [history]);
 
-  const educationHistory = useMemo(
-    () =>
-      getHistoryProps(
-        profile ? profile.education_histories : [],
-        HISTORY_TYPE.education
-      ),
-    [profile]
-  );
-
-  const workHistory = useMemo(
-    () =>
-      getHistoryProps(
-        profile ? profile.work_experiences : [],
-        HISTORY_TYPE.work
-      ),
-    [profile]
-  );
-
   const {languages, keywords} = useMemo(()=>{
     if (profile && profile.competences) {
       const competences = getCompetencesWithCategoryNames(competenceCategories, profile.competences);
@@ -97,7 +51,7 @@ function ProfilePage({ id, onSave }) {
     return <h1>404</h1>;
   }
 
-  const profileProps = {...profile, educationHistory, workHistory, languages, keywords};
+  const profileProps = {...profile, languages, keywords};
 
   if(edit) {
     return (
