+++ web/src/pages/ProfilePage/ProfilePageView.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useMemo } from "react";
 
 import { Button } from "@mui/material";
 import { styled } from "@mui/system";
@@ -78,11 +78,55 @@ const Education = styled("div")`
   }
 `;
 
+const HISTORY_TYPE = {
+  education: "Education",
+  work: "Work",
+};
+/**
+ * A function that produces the props for using HistoryList component
+ *
+ * @param {Array<object>} historyItems - Array of history items
+ * @param {*} type - type of history items
+ * @returns {object}
+ */
+ const getHistoryProps = (historyItems = [], type) => {
+  return {
+    title: `${type} History`,
+    noItemDescription: `No ${type} History Provided`,
+    historyItems: historyItems.map((historyItem) => ({
+      id: historyItem.id,
+      organisation:
+        historyItem[type === HISTORY_TYPE.education ? "school" : "company"],
+      title:
+        historyItem[type === HISTORY_TYPE.education ? "degree" : "position"],
+      description: historyItem.description,
+      start_date: historyItem.start_date,
+      end_date: historyItem.end_date,
+    })),
+  };
+};
 const image_url = "https://images.pexels.com/photos/6386956/pexels-photo-6386956.jpeg?auto=compress&cs=tinysrgb&dpr=1&w=500";
 
 function ProfilePageView({ profile, onEditClick }) {
-  const {languages, keywords, workHistory, educationHistory} = profile;
+  const { languages, keywords } = profile;
 
+  const educationHistory = useMemo(
+    () =>
+      getHistoryProps(
+        profile ? profile.education_histories : [],
+        HISTORY_TYPE.education
+      ),
+    [profile]
+  );
+
+  const workHistory = useMemo(
+    () =>
+      getHistoryProps(
+        profile ? profile.work_experiences : [],
+        HISTORY_TYPE.work
+      ),
+    [profile]
+  );
   return (
     <Page header={
       profile &&
@@ -140,13 +184,13 @@ function ProfilePageView({ profile, onEditClick }) {
           noItemDescription={educationHistory.noItemDescription}
         />
       </Education>
-      
+
       <ActionButtonContainer>
         <Button
-            fullWidth
-            variant="contained"
-            color="primary"
-            onClick={onEditClick}
+          fullWidth
+          variant="contained"
+          color="primary"
+          onClick={onEditClick}
         >
           Edit
         </Button>
